+++ crates/cli/src/commands/setup.rs
@@ -8,12 +8,14 @@ use crate::{
     util::{check_private_keys, find_insufficient_balances, fund_accounts, load_seedfile},
     LATENCY_HIST as HIST, PROM,
 };
+use contender_core::util::get_block_time;
 use contender_core::{
-    agent_controller::{AgentStore, SignerStore},
-    generator::RandSeed,
+    generator::{
+        agent_pools::{AgentPools, AgentSpec},
+        RandSeed,
+    },
     test_scenario::{TestScenario, TestScenarioParams},
 };
-use contender_core::{generator::PlanConfig, util::get_block_time};
 use contender_engine_provider::DEFAULT_BLOCK_TIME;
 use contender_testfile::TestConfig;
 use std::{
@@ -73,25 +75,12 @@ pub async fn setup(
         return Err(SetupError::insufficient_funds(broke_accounts).into());
     }
 
-    // load agents from setup and create pools
-    let from_pool_declarations =
-        [testconfig.get_setup_pools(), testconfig.get_create_pools()].concat();
-
-    // create agents for each from_pool declaration
-    let mut agents = AgentStore::new();
-
-    // If override_senders is true, we don't need to create agents for pools
-    // since all transactions will use the primary signer
-    if !override_senders {
-        for from_pool in &from_pool_declarations {
-            if agents.has_agent(from_pool) {
-                continue;
-            }
-
-            let agent = SignerStore::new(accounts_per_agent, &args.seed, from_pool);
-            agents.add_agent(from_pool, agent);
-        }
-    }
+    // create agents for each from_pool declaration in setup and create pools
+    let agent_spec = AgentSpec::default()
+        .create_accounts(accounts_per_agent)
+        .setup_accounts(accounts_per_agent)
+        .spam_accounts(0);
+    let agents = testconfig.build_agent_store(&args.seed, agent_spec.clone());
 
     // user-provided signers must be pre-funded
     let admin_signer = args.eth_json_rpc_args.primary_signer();
@@ -113,7 +102,7 @@ pub async fn setup(
         rpc_url: args.eth_json_rpc_args.rpc_url,
         builder_rpc_url: None,
         signers: user_signers_with_defaults,
-        agent_store: agents,
+        agent_spec,
         tx_type: tx_type.into(),
         bundle_type: bundle_type.into(),
         pending_tx_timeout_secs: 12,
@@ -122,6 +111,7 @@ pub async fn setup(
         rpc_batch_size: 0,
         gas_price: None,
     };
+
     let mut scenario = TestScenario::new(
         testconfig.to_owned(),
         db.clone().into(),